+++ src/app/mat-select-search/mat-select-search-icon.directive.ts
@@ -1,0 +1,13 @@
+import { Directive } from '@angular/core';
+
+/**
+ * Directive for providing a custom search-icon.
+ * e.g.
+ * <ngx-mat-select-search [formControl]="bankFilterCtrl">
+ *   <mat-icon ngxMatSelectSearchIcon>search</mat-icon>
+ * </ngx-mat-select-search>
+ */
+@Directive({
+    selector: '[ngxMatSelectSearchIcon]'
+})
+export class MatSelectSearchIconDirective {}