+++ src/app/mat-select-search/mat-select-search.component.ts
@@ -32,6 +32,8 @@ import { delay, filter, map, startWith, switchMap, take, takeUntil, tap } from '
 import { MatSelectSearchClearDirective } from './mat-select-search-clear.directive';
 import { configurableDefaultOptions, MAT_SELECTSEARCH_DEFAULT_OPTIONS, MatSelectSearchOptions } from './default-options';
 import { MatSelectNoEntriesFoundDirective } from './mat-select-no-entries-found.directive';
+import { MatSelectSearchIconDirective } from './mat-select-search-icon.directive';
+
 
 /**
  * Component providing an input field for searching MatSelect options.
@@ -214,6 +216,9 @@ export class MatSelectSearchComponent implements OnInit, OnDestroy, ControlValue
 
   /** Reference to custom no entries found element */
   @ContentChild(MatSelectNoEntriesFoundDirective) noEntriesFound: MatSelectNoEntriesFoundDirective;
+
+  /** Reference to custom search icon */
+  @ContentChild(MatSelectSearchIconDirective, { static: false }) searchIcon: MatSelectSearchIconDirective;
 
   /** Current search value */
   get value(): string {