+++ src/app/mat-select-search/ngx-mat-select-search.module.ts
@@ -15,6 +15,7 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { CommonModule } from '@angular/common';
 
 import { MatSelectSearchClearDirective } from './mat-select-search-clear.directive';
+import { MatSelectSearchIconDirective } from './mat-select-search-icon.directive';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSelectNoEntriesFoundDirective } from './mat-select-no-entries-found.directive';
 import { MatDividerModule } from '@angular/material/divider';
@@ -37,11 +38,13 @@ export { MatSelectNoEntriesFoundDirective };
   declarations: [
     MatSelectSearchComponent,
     MatSelectSearchClearDirective,
+    MatSelectSearchIconDirective,
     MatSelectNoEntriesFoundDirective
   ],
   exports: [
     MatSelectSearchComponent,
     MatSelectSearchClearDirective,
+    MatSelectSearchIconDirective,
     MatSelectNoEntriesFoundDirective
   ]
 })