+++ src/app/mat-select-search/public_api.ts
@@ -1,3 +1,4 @@
 export * from './mat-select-search.component';
+export * from './mat-select-search-icon.directive';
 export * from './ngx-mat-select-search.module';
 export * from './default-options';